Student validation (#74)
* deletion of select lib Select components

* new select component

* tela de criar projeto consertada

* tela de criar projeto consertada

* update user in actvity

* resolving deploying conflicts

* user can enroll in view project activity

* Basic validations for create project form

* light fixes about project

* feat: Add export functionality for approved students

The code changes in this commit add a new function `handleExportApprovedStudents` to the `ViewEnrolledStudents` component. This function allows exporting a CSV file containing the details of approved students for a project. The function filters the list of applicants to include only those with a `status` of `true`, and then generates a CSV file with the student's name, email, and RA (split from the email address). The generated CSV file is then downloaded by creating a temporary URL and clicking on a dynamically created anchor element.

Based on the recent commits and repository history, the commit message follows the established convention of starting with a type prefix (`feat` for a new feature) and provides a clear and concise description of the changes made.

* student can't access admin functions, change project to coil

---------

Co-authored-by: luiz-felipe-b <lfbalaminute@hotmail.com>
Co-authored-by: Iris <iris.melero@outlook.com>
Co-authored-by: FelipeCarillo <felipecarillo@outlook.com>